Create best_params.py
with different cluster ranging from 2 to any specific defined is taken.
total keeps a track of all the data points in the particular cluster.
clusters keep a track of the target points in that particular clusters.
per is percentage of target points in a particular cluster.
support is the ratio of how good a particular cluster is.

best_params so generated is a list of dictionaries of all the aforementioned parameters.
support has been sorted while appending. hence, we can evaluate the results quickly with respect to support.